requirements.txt
streamlit
google-generativeai
python-dotenv
pymupdf
speechrecognition
pydub
requests
pytest
simpleaudio